Move key_or_iv into a helpers module
