+++ src/transformation/transform_stream_cipher.rs
@@ -10,7 +10,7 @@ use simple_error::{SimpleResult, bail};
 use serde::{Serialize, Deserialize};
 
 use crate::transformation::TransformerTrait;
-use crate::transformation::key_or_iv::KeyOrIV;
+use crate::transformation::helpers::key_or_iv::KeyOrIV;
 
 /// Which stream cipher should we use?
 #[allow(non_camel_case_types)]
